# Optimizations: Function Order, Removal of Unused Logic, and Struct