--- a/foundry.toml
+++ b/foundry.toml
@@ -1,9 +1,14 @@
 [profile.default]
-src = 'src'
-out = 'out'
-libs = ['lib']
+auto_detect_solc = false
+bytecode_hash = "none"
+evm_version = "paris"           # See https://www.evmdiff.com/features?name=PUSH0&kind=opcode
+fuzz = { runs = 1_000 }
+gas_reports = ["*"]
 optimizer = true
-optimizer_runs = 10
-via_ir = false
-
+optimizer_runs = 200
+out = "out"
+script = "script"
+solc = "0.8.24"
+src = "src"
+test = "test"
 # See more config options https://github.com/foundry-rs/foundry/tree/master/config