--- a/foundry.toml
+++ b/foundry.toml
@@ -1,11 +1,9 @@
 [profile.default]
-auto_detect_solc = false
 bytecode_hash = "none"
-evm_version = "paris"           # See https://www.evmdiff.com/features?name=PUSH0&kind=opcode
 fuzz = { runs = 1_000 }
 gas_reports = ["*"]
 optimizer = true
-optimizer_runs = 200
+optimizer_runs = 16
 out = "out"
 script = "script"
 solc = "0.8.24"