… GasContract.svg
… classDiagram.svg
… src/Ownable.sol