Summary:
This pull request introduces several optimizations aimed at improving the efficiency and readability of the GasContract smart contract. The changes include adjusting the order of functions for better logical grouping, removing the checkForAdmin logic which was deemed unnecessary, and eliminating the ImportantStruct which was not being utilized.

Details:

- Change Function Order: Functions have been reordered to group similar functionalities together, which enhances the contract's readability.
- Remove checkForAdmin Logic: Given its limited use in the current contract logic, this part has been removed to simplify the codebase.
- Remove ImportantStruct: This struct was not in use and has been removed to clean up the code.
- Remove Ownable.sol
- Remove transfer event
- Add Makefile to util feature, such as run test , run sol2uml command.

## How much reduce Deployment gas Cost?
This change reduce deployment gas cost from 650329 to 359883 in my local environment.

## Before
```
❯ forge test --gas-report --watch
[⠢] Compiling...
No files changed, compilation skipped

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 362)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 25394, ~: 25403)
[PASS] testBalanceOf() (gas: 12181)
[PASS] testCheckForAdmin() (gas: 19630)
[PASS] testGetPaymentHistory() (gas: 428)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 418, ~: 418)
[PASS] testGetTradingMode() (gas: 670)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 45775, ~: 47726)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 122871, ~: 123304)
[PASS] test_admins() (gas: 35798)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 27431, ~: 27596)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 28504, ~: 28640)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 23544, ~: 23544)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 116745, ~: 117112)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 31302, ~: 31420)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 116698, ~: 117059)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 403.99ms (1.17s CPU time)
| src/Gas.sol:GasContract contract |                 |       |        |       |         |
|----------------------------------|-----------------|-------|--------|-------|---------|
| Deployment Cost                  | Deployment Size |       |        |       |         |
| 650329                           | 3039            |       |        |       |         |
| Function Name                    | min             | avg   | median | max   | # calls |
| addToWhitelist                   | 11725           | 13048 | 13002  | 13828 | 8       |
| administrators                   | 2633            | 2633  | 2633   | 2633  | 5       |
| balanceOf                        | 570             | 2070  | 2570   | 2570  | 8       |
| balances                         | 508             | 1008  | 508    | 2508  | 4       |
| checkForAdmin                    | 11730           | 11730 | 11730  | 11730 | 1       |
| getPaymentStatus                 | 658             | 658   | 658    | 658   | 1       |
| transfer                         | 25405           | 26905 | 27405  | 27405 | 4       |
| whiteTransfer                    | 69028           | 70361 | 71028  | 71028 | 3       |
| whitelist                        | 640             | 640   | 640    | 640   | 2       |




Ran 1 test suite in 406.38ms (403.99ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```
## After
```
❯ make test
forge test --gas-report
[⠊] Compiling...
[⠰] Compiling 2 files with 0.8.24
[⠔] Solc 0.8.24 finished in 1.41s
Compiler run successful with warnings:
Warning (5667): Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> src/Gas.sol:10:17:
   |
10 |     constructor(address[] memory _admins, uint256 _totalSupply) {
   |                 ^^^^^^^^^^^^^^^^^^^^^^^^

Warning (5667): Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> src/Gas.sol:10:43:
   |
10 |     constructor(address[] memory _admins, uint256 _totalSupply) {
   |                                           ^^^^^^^^^^^^^^^^^^^^

Warning (5667): Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> src/Gas.sol:29:9:
   |
29 |         string calldata _name
   |         ^^^^^^^^^^^^^^^^^^^^^

Warning (5667): Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> src/Gas.sol:44:9:
   |
44 |         address sender
   |         ^^^^^^^^^^^^^^

Warning (5667): Unused function parameter. Remove or comment out the variable name to silence this warning.
  --> src/Gas.sol:57:24:
   |
57 |     function whitelist(address addr) external pure returns (uint256) {
   |                        ^^^^^^^^^^^^


Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 362)
[PASS] testAddToWhitelist(address,uint256) (runs: 1000, μ: 11914, ~: 11914)
[PASS] testBalanceOf() (gas: 12181)
[PASS] testCheckForAdmin() (gas: 8371)
[PASS] testGetPaymentHistory() (gas: 428)
[PASS] testGetPaymentStatus(address) (runs: 1000, μ: 418, ~: 418)
[PASS] testGetTradingMode() (gas: 670)
[PASS] testTransfer(uint256,address) (runs: 1000, μ: 44686, ~: 46330)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 1000, μ: 85181, ~: 85584)
[PASS] test_admins() (gas: 35578)
[PASS] test_onlyOwner(address,uint256) (runs: 1000, μ: 13962, ~: 14107)
[PASS] test_tiers(address,uint256) (runs: 1000, μ: 17207, ~: 17348)
[PASS] test_tiersReverts(address,uint256) (runs: 1000, μ: 12169, ~: 12169)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 1000, μ: 79534, ~: 79923)
[PASS] test_whitelistEvents(address,uint256) (runs: 1000, μ: 19988, ~: 20128)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 1000, μ: 79303, ~: 79636)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 1.35s (3.96s CPU time)
| src/Gas.sol:GasContract contract |                 |       |        |       |         |
|----------------------------------|-----------------|-------|--------|-------|---------|
| Deployment Cost                  | Deployment Size |       |        |       |         |
| 359883                           | 1902            |       |        |       |         |
| Function Name                    | min             | avg   | median | max   | # calls |
| addToWhitelist                   | 339             | 1197  | 1710   | 1710  | 8       |
| administrators                   | 2589            | 2589  | 2589   | 2589  | 5       |
| balanceOf                        | 570             | 2070  | 2570   | 2570  | 8       |
| balances                         | 508             | 1008  | 508    | 2508  | 4       |
| checkForAdmin                    | 471             | 471   | 471    | 471   | 1       |
| getPaymentStatus                 | 554             | 554   | 554    | 554   | 1       |
| transfer                         | 24009           | 25509 | 26009  | 26009 | 4       |
| whiteTransfer                    | 44412           | 45745 | 46412  | 46412 | 3       |
| whitelist                        | 432             | 432   | 432    | 432   | 2       |

Ran 1 test suite in 1.35s (1.35s CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```